+++ svelte/demos/cases/BasicInit.svelte
@@ -12,8 +12,8 @@
 
 	let message = $state("");
 	function clicked(ev) {
-		const action = ev.action;
-		message = action ? `clicked on ${action.id}` : "closed";
+		const option = ev.option;
+		message = option ? `clicked on ${option.id}` : "closed";
 		menu1 = menu2 = menu3 = null;
 	}
 </script>