+++ svelte/src/components/DropDownMenu.svelte
@@ -4,17 +4,17 @@
 
 	let { options, at = "bottom", css = "", children, onclick } = $props();
 
-	export const handler = ev => {
+	export function show(ev) {
 		parent = ev.target;
 		ev.preventDefault();
-	};
+	}
 
 	var parent = $state(null);
 	function onClick(ev) {
 		parent = null;
 		onclick && onclick(ev);
 	}
-	function show(ev) {
+	function showAt(ev) {
 		let target = ev.target;
 		while (!target.dataset.menuIgnore) {
 			parent = target;
@@ -25,7 +25,7 @@
 
 <!-- svelte-ignore a11y_click_events_have_key_events -->
 <!-- svelte-ignore a11y_no_static_element_interactions -->
-<span onclick={show} data-menu-ignore="true">
+<span onclick={showAt} data-menu-ignore="true">
 	{@render children?.()}
 </span>
 {#if parent}